move 'ProductTypeIconMapping' to common project model
